refactor(extensions): separate lib extension code
