+++ assets/express/src/extensions/knex/register.ts
@@ -1,0 +1,3 @@
+import { register } from './index';
+
+register();
