+++ assets/express/src/extensions/zod/index.ts
@@ -1,0 +1,12 @@
+import z from 'zod';
+import { extendZodWithOpenApi } from '@asteasolutions/zod-to-openapi';
+
+declare module '@asteasolutions/zod-to-openapi' {
+  interface ZodOpenAPIMetadata {
+    schemaName: string;
+  }
+}
+
+export const register = () => {
+  extendZodWithOpenApi(z);
+};
