+++ assets/express/src/extensions/zod/register.ts
@@ -1,0 +1,3 @@
+import { register } from './index';
+
+register();
